+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyleableInfo.java
@@ -30,6 +30,9 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import com.formdev.flatlaf.icons.*;
+import com.formdev.flatlaf.ui.TestFlatStyling.CustomCheckBoxIcon;
+import com.formdev.flatlaf.ui.TestFlatStyling.CustomIcon;
+import com.formdev.flatlaf.ui.TestFlatStyling.CustomRadioButtonIcon;
 
 /**
  * @author Karl Tauber
@@ -144,7 +147,12 @@ public class TestFlatStyleableInfo
 
 	@Test
 	void checkBox() {
-		JCheckBox c = new JCheckBox();
+		checkBox( new JCheckBox() );
+		checkBox( new JCheckBox( new CustomIcon() ) );
+		checkBox( new JCheckBox( new CustomCheckBoxIcon() ) );
+	}
+
+	private void checkBox( JCheckBox c ) {
 		FlatCheckBoxUI ui = (FlatCheckBoxUI) c.getUI();
 
 		assertTrue( ui.getDefaultIcon() instanceof FlatCheckBoxIcon );
@@ -152,6 +160,11 @@ public class TestFlatStyleableInfo
 		// FlatCheckBoxUI extends FlatRadioButtonUI
 		Map<String, Class<?>> expected = new LinkedHashMap<>();
 		radioButton( expected );
+
+		// remove "icon." keys if check box has custom icon
+		Icon icon = c.getIcon();
+		if( icon != null && !(icon instanceof FlatCheckBoxIcon) )
+			expected.keySet().removeIf( key -> key.startsWith( "icon." ) );
 
 		assertMapEquals( expected, ui.getStyleableInfos( c ) );
 	}
@@ -492,7 +505,12 @@ public class TestFlatStyleableInfo
 
 	@Test
 	void radioButton() {
-		JRadioButton c = new JRadioButton();
+		radioButton( new JRadioButton() );
+		radioButton( new JRadioButton( new CustomIcon() ) );
+		radioButton( new JRadioButton( new CustomRadioButtonIcon() ) );
+	}
+
+	private void radioButton( JRadioButton c ) {
 		FlatRadioButtonUI ui = (FlatRadioButtonUI) c.getUI();
 
 		assertTrue( ui.getDefaultIcon() instanceof FlatRadioButtonIcon );
@@ -503,6 +521,11 @@ public class TestFlatStyleableInfo
 		expectedMap( expected,
 			"icon.centerDiameter", float.class
 		);
+
+		// remove "icon." keys if radio button has custom icon
+		Icon icon = c.getIcon();
+		if( icon != null && !(icon instanceof FlatRadioButtonIcon) )
+			expected.keySet().removeIf( key -> key.startsWith( "icon." ) );
 
 		assertMapEquals( expected, ui.getStyleableInfos( c ) );
 	}
@@ -1113,6 +1136,16 @@ public class TestFlatStyleableInfo
 	}
 
 	@Test
+	void flatScrollPaneBorder() {
+		FlatScrollPaneBorder border = new FlatScrollPaneBorder();
+
+		Map<String, Class<?>> expected = new LinkedHashMap<>();
+		flatScrollPaneBorder( expected );
+
+		assertMapEquals( expected, border.getStyleableInfos() );
+	}
+
+	@Test
 	void flatTextBorder() {
 		FlatTextBorder border = new FlatTextBorder();
 
@@ -1291,6 +1324,58 @@ public class TestFlatStyleableInfo
 			"pressedBackground", Color.class,
 			"questionMarkColor", Color.class,
 			"disabledQuestionMarkColor", Color.class
+		);
+
+		assertMapEquals( expected, icon.getStyleableInfos() );
+	}
+
+	@Test
+	void flatClearIcon() {
+		FlatClearIcon icon = new FlatClearIcon();
+
+		Map<String, Class<?>> expected = expectedMap(
+			"clearIconColor", Color.class,
+			"clearIconHoverColor", Color.class,
+			"clearIconPressedColor", Color.class
+		);
+
+		assertMapEquals( expected, icon.getStyleableInfos() );
+	}
+
+	@Test
+	void flatSearchIcon() {
+		FlatSearchIcon icon = new FlatSearchIcon();
+
+		Map<String, Class<?>> expected = new LinkedHashMap<>();
+		flatSearchIcon( expected );
+
+		assertMapEquals( expected, icon.getStyleableInfos() );
+	}
+
+	@Test
+	void flatSearchWithHistoryIcon() {
+		FlatSearchWithHistoryIcon icon = new FlatSearchWithHistoryIcon();
+
+		Map<String, Class<?>> expected = new LinkedHashMap<>();
+		flatSearchIcon( expected );
+
+		assertMapEquals( expected, icon.getStyleableInfos() );
+	}
+
+	private void flatSearchIcon( Map<String, Class<?>> expected ) {
+		expectedMap( expected,
+			"searchIconColor", Color.class,
+			"searchIconHoverColor", Color.class,
+			"searchIconPressedColor", Color.class
+		);
+	}
+
+	@Test
+	void flatCapsLockIcon() {
+		FlatCapsLockIcon icon = new FlatCapsLockIcon();
+
+		Map<String, Class<?>> expected = expectedMap(
+			"capsLockIconColor", Color.class
 		);
 
 		assertMapEquals( expected, icon.getStyleableInfos() );