+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyling.java
@@ -19,10 +19,15 @@ package com.formdev.flatlaf.ui;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.awt.Color;
+import java.awt.Component;
 import java.awt.Dimension;
+import java.awt.Graphics;
 import java.awt.Insets;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Set;
 import java.util.function.Consumer;
 import javax.swing.*;
 import javax.swing.table.JTableHeader;
@@ -31,6 +36,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import com.formdev.flatlaf.FlatLaf;
 import com.formdev.flatlaf.icons.*;
+import com.formdev.flatlaf.ui.FlatStylingSupport.UnknownStyleException;
 import com.formdev.flatlaf.util.ColorFunctions;
 
 /**
@@ -47,6 +53,14 @@ public class TestFlatStyling
 		FlatLaf.setGlobalExtraDefaults( globalExtraDefaults );
 
 		TestUtils.setup( false );
+
+		Set<String> excludes = new HashSet<>();
+		Collections.addAll( excludes,
+			"parse", "parseIfFunction", "parseColorFunctions",
+			"parseReferences", "parseVariables", "parseRecursiveVariables",
+			"enumField", "enumProperty", "enumUIDefaults" );
+		TestUtils.checkImplementedTests( excludes, TestFlatStyling.class,
+			TestFlatStyleableValue.class, TestFlatStyleableInfo.class );
 	}
 
 	@AfterAll
@@ -294,13 +308,22 @@ public class TestFlatStyling
 
 	@Test
 	void checkBox() {
-		JCheckBox c = new JCheckBox();
+		checkBox( new JCheckBox() );
+		checkBox( new JCheckBox( new CustomIcon() ) );
+		checkBox( new JCheckBox( new CustomCheckBoxIcon() ) );
+	}
+
+	private void checkBox( JCheckBox c ) {
 		FlatCheckBoxUI ui = (FlatCheckBoxUI) c.getUI();
 
 		assertTrue( ui.getDefaultIcon() instanceof FlatCheckBoxIcon );
 
 		// FlatCheckBoxUI extends FlatRadioButtonUI
 		radioButton( ui, c );
+
+		// necessary to clear FlatRadioButtonUI.oldStyleValues because
+		// ui.applyStyle(...) operates on shared instance
+		ui.uninstallUI( c );
 	}
 
 	@Test
@@ -651,14 +674,24 @@ public class TestFlatStyling
 
 	@Test
 	void radioButton() {
-		JRadioButton c = new JRadioButton();
+		radioButton( new JRadioButton() );
+		radioButton( new JRadioButton( new CustomIcon() ) );
+		radioButton( new JRadioButton( new CustomRadioButtonIcon() ) );
+	}
+
+	private void radioButton( JRadioButton c ) {
 		FlatRadioButtonUI ui = (FlatRadioButtonUI) c.getUI();
 
 		assertTrue( ui.getDefaultIcon() instanceof FlatRadioButtonIcon );
 
 		radioButton( ui, c );
 
-		ui.applyStyle( c, "icon.centerDiameter: 8" );
+		if( !(c.getIcon() instanceof CustomIcon) )
+			ui.applyStyle( c, "icon.centerDiameter: 8" );
+
+		// necessary to clear FlatRadioButtonUI.oldStyleValues because
+		// ui.applyStyle(...) operates on shared instance
+		ui.uninstallUI( c );
 	}
 
 	private void radioButton( FlatRadioButtonUI ui, AbstractButton b ) {
@@ -675,6 +708,16 @@ public class TestFlatStyling
 		ui.applyStyle( b, "iconTextGap: 4" );
 
 		//---- icon ----
+
+		if( b.getIcon() instanceof CustomIcon ) {
+			try {
+				ui.applyStyle( b, "icon.focusWidth: 1.5" );
+				assertTrue( false );
+			} catch( UnknownStyleException ex ) {
+				assertEquals( new UnknownStyleException( "icon.focusWidth" ).getMessage(), ex.getMessage() );
+			}
+			return;
+		}
 
 		ui.applyStyle( b, "icon.focusWidth: 1.5" );
 		ui.applyStyle( b, "icon.focusColor: #fff" );
@@ -1334,6 +1377,16 @@ public class TestFlatStyling
 	}
 
 	@Test
+	void flatScrollPaneBorder() {
+		FlatScrollPaneBorder border = new FlatScrollPaneBorder();
+
+		// FlatScrollPaneBorder extends FlatBorder
+		flatBorder( border );
+
+		border.applyStyleProperty( "arc", 6 );
+	}
+
+	@Test
 	void flatTextBorder() {
 		FlatTextBorder border = new FlatTextBorder();
 
@@ -1532,6 +1585,13 @@ public class TestFlatStyling
 		icon.applyStyleProperty( "searchIconPressedColor", Color.WHITE );
 	}
 
+	@Test
+	void flatCapsLockIcon() {
+		FlatCapsLockIcon icon = new FlatCapsLockIcon();
+
+		icon.applyStyleProperty( "capsLockIconColor", Color.WHITE );
+	}
+
 	//---- enums --------------------------------------------------------------
 
 	enum SomeEnum { enumValue1, enumValue2 }
@@ -1564,5 +1624,35 @@ public class TestFlatStyling
 
 		UIManager.put( "test.enum", null );
 		assertEquals( SomeEnum.enumValue1, FlatUIUtils.getUIEnum( "test.enum", SomeEnum.class, SomeEnum.enumValue1 ) );
+	}
+
+	//---- class CustomIcon ---------------------------------------------------
+
+	static class CustomIcon
+		implements Icon
+	{
+		@Override public void paintIcon( Component c, Graphics g, int x, int y ) {}
+		@Override public int getIconWidth() { return 1; }
+		@Override public int getIconHeight() { return 1; }
+	}
+
+	//---- class CustomCheckBoxIcon ----------------------------------------
+
+	static class CustomCheckBoxIcon
+		extends FlatCheckBoxIcon
+	{
+		CustomCheckBoxIcon() {
+			background = Color.green;
+		}
+	}
+
+	//---- class CustomRadioButtonIcon ----------------------------------------
+
+	static class CustomRadioButtonIcon
+		extends FlatRadioButtonIcon
+	{
+		CustomRadioButtonIcon() {
+			background = Color.green;
+		}
 	}
 }